move tasks test to dedicated folder
